+++ src/eBayEnterprise/RetailOrderManagement/Payload/BidirectionalMessageConfigMap.php
@@ -107,6 +107,10 @@ return call_user_func(function () {
         'request' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\PaymentSettlementRequest',
         'reply' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\PaymentSettlementReply',
     ];
+    $map['payments/auth/cancel'] = [
+        'request' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\PaymentAuthCancelRequest',
+        'reply' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\PaymentAuthCancelReply',
+    ];
     $map['payments/funds/confirm'] = [
         'request' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\ConfirmFundsRequest',
         'reply' => '\eBayEnterprise\RetailOrderManagement\Payload\Payment\ConfirmFundsReply',