+++ src/eBayEnterprise/RetailOrderManagement/Payload/Payment/ConfirmFundsReply.php
@@ -26,6 +26,7 @@ use Psr\Log\NullLogger;
 class ConfirmFundsReply implements IConfirmFundsReply
 {
     const SUCCESS = 'Success';
+    const TIMEOUT = 'Timeout';
 
     use TTopLevelPayload, TPaymentContext;
 
@@ -53,7 +54,7 @@ class ConfirmFundsReply implements IConfirmFundsReply
         $this->parentPayload = $parentPayload;
 
         $this->extractionPaths = [
-            'orderId' => 'string(x:PaymentContext/x:OrderId)',
+            'orderId' => 'string(x:PaymentContext/x:OrderId|x:PaymentContextBase/x:OrderId)',
             'cardNumber' =>
                 'string(x:PaymentContext/x:EncryptedPaymentAccountUniqueId|x:PaymentContext/x:PaymentAccountUniqueId)',
             'fundsAvailable' => 'string(x:FundsAvailable)',
@@ -68,6 +69,16 @@ class ConfirmFundsReply implements IConfirmFundsReply
     public function isSuccess()
     {
         return ($this->getFundsAvailable() === static::SUCCESS);
+    }
+
+    /**
+     * Did upstream systems report a timeout?
+     *
+     * @return bool
+     */
+    public function isTimeout()
+    {
+        return ($this->getFundsAvailable() === static::TIMEOUT);
     }
 
     /**
@@ -98,7 +109,10 @@ class ConfirmFundsReply implements IConfirmFundsReply
      */
     protected function serializeContents()
     {
-        return $this->serializePaymentContext() .
+        $paymentContext = $this->getCardNumber() ?
+            $this->serializePaymentContext() :
+            $this->serializePaymentContextBase();
+        return $paymentContext .
         "<FundsAvailable>{$this->xmlEncode($this->getFundsAvailable())}</FundsAvailable>";
     }